Move the A* implementation into its own header file.
